Add files via upload
Get a to-do list and a timer.

Set your timer for 25 minutes, and focus on a single task until the timer rings.

When your session ends, mark off one pomodoro and record what you completed.

Then enjoy a five-minute break.

After your task is complete. calculate the total time and display in the complete task section with task name.